Initial BLEautoconnect version upload
Based on Gameboy Enhacned Firmware version 0.77, it includes all supergame boy colour mapping automatically applied, if no super gameboy colour data is present it will revert to the gameboy pocket palette. This is the first release, palette control is planned for future updates to match the options available in the full emulator project.